Moved activity page child components into organized sub-directories and updated import paths.
