+++ src/components/activity/Filters.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef, useState } from "react";
 import { connect } from "react-redux";
 import { quakeFetch, updateSearchParams } from "../../actions";
-import SearchBar from "./SearchBar";
+import SearchBar from "./search/SearchBar";
 import useDarkMode from "../../customHooks/useDarkMode";
 
 function Filters({
